Share the analytics SQL helpers between the raw and rollup reads
