+++ apps/web/src/server/analytics-core.ts
@@ -26,15 +26,14 @@ import {
 	getPromptsSummary,
 	getVisibilityDailyAggregate,
 } from "@/lib/analytics-read";
+import { API_PROVIDER_IDS, isCalendarDay } from "@/lib/analytics-sql";
 import { generateDateRange } from "@/lib/chart-utils";
 import { computeFanoutAnalysis, type FanoutAnalysis, type FanoutLimitOverrides } from "@/lib/fanout-analysis";
 import {
-	API_PROVIDER_IDS,
 	getFanoutBreakdown,
 	getFanoutModelTotals,
 	getFanoutPromptTotals,
 	getPromptsFirstEvaluatedAt,
-	isCalendarDay,
 } from "@/lib/postgres-read";
 import { computeShareOfVoice, shareOfVoiceLeaderboardLVCF, shareOfVoiceTimeSeriesLVCF } from "@/lib/visibility-stats";
 import { resolveFilteredPrompts } from "@/server/prompt-resolution";
